backend/requirements.txt
Django==4.1.3
PyPDF2==3.0.1
google-generativeai==0.8.3